Remove wasm-dependencies from flnode
