+++ Makefile
@@ -1,5 +1,5 @@
 update_version:
-	echo "pub const VERSION_STRING: &str = \"$$( date +%Y-%m-%d_%H:%M )::$$( git rev-parse --short HEAD )\";" > common/src/node/version.rs
+	echo "pub const VERSION_STRING: &str = \"$$( date +%Y-%m-%d_%H:%M )::$$( git rev-parse --short HEAD )\";" > flnode/src/node/version.rs
 
 kill_bg:
 	pkill -f signal &
@@ -8,10 +8,10 @@ kill_bg:
 build_local:
 	cargo build
 	make -C web clean build_local 
-	make -C cli/flnode build_local
+	make -C cli/fledger build_local
 
 serve_local: kill_bg build_local
 	cargo run --bin signal &
 	make -C web serve &
-	make -C cli/flnode run2
+	make -C cli/fledger run2
 	open localhost:8080
