+++ flnet/Cargo.toml
@@ -7,28 +7,21 @@ edition = "2021"
 
 [dependencies]
 flutils = {path = "../flutils"}
-flmodules = {path = "../flmodules"}
-transitive_from = "0.1.3"
-async-trait = "0.1.51"
-rand = "0.7"
-serde = { version = "1.0.127", features = ["derive"] }
-serde_derive = "1.0.127"
-serde_json = "1.0.66"
-serde_yaml = "0.8"
-toml = "0.5.8"
 
 # The "js" is to make sure it also works for the flnode and web crates
 getrandom = { version = "0.2.3", features = ["js"] }
-names = { path = "../vendor/names" }
+async-trait = "0.1.51"
+rand = "0.7.3"
+serde = { version = "1.0.127", features = ["derive"] }
+serde_derive = "1.0.127"
+serde_json = "1.0.66"
+toml = "0.5.8"
+names = "0.12.0"
 thiserror = "1.0.26"
-
 futures = "0.3.16"
 log = "0.4.14"
-sha2 = "0.9.5"
-
+sha2 = "0.10.1"
 ed25519-dalek = { version = "1.0.1", features = ["serde"] }
-timer = "0.2"
-itertools = "0.10.3"
 
 [dependencies.web-sys]
 features = [
