+++ flnode/Cargo.toml
@@ -10,9 +10,10 @@ version = "0.1.0"
 flnet = {path="../flnet"}
 flutils = {path = "../flutils"}
 flmodules = {path = "../flmodules"}
+
 transitive_from = "0.1.3"
 async-trait = "0.1.51"
-rand = "0.7"
+rand = "0.8.4"
 serde = { version = "1.0.127", features = ["derive"] }
 serde_derive = "1.0.127"
 serde_json = "1.0.66"
@@ -21,55 +22,17 @@ toml = "0.5.8"
 
 # The "js" is to make sure it also works for the flnode and web crates
 getrandom = { version = "0.2.3", features = ["js"] }
-names = { path = "../vendor/names" }
+names = "0.12.0"
 thiserror = "1.0.26"
 
 futures = "0.3.16"
 log = "0.4.14"
-sha2 = "0.9.5"
+sha2 = "0.10.1"
 
 ed25519-dalek = { version = "1.0.1", features = ["serde"] }
 timer = "0.2"
 itertools = "0.10.3"
-
-[dependencies.web-sys]
-features = [
-  "Storage",
-  "Request",
-  "RequestInit",
-  "RequestMode",
-  "Response",
-  "MessageEvent",
-  "RtcPeerConnection",
-  "RtcSignalingState",
-  "RtcSdpType",
-  "RtcSessionDescriptionInit",
-  "RtcPeerConnectionIceEvent",
-  "RtcIceCandidate",
-  "RtcIceCandidateInit",
-  "RtcIceConnectionState",
-  "RtcIceGatheringState",
-  "RtcDataChannel",
-  "RtcDataChannelEvent",
-  "RtcDataChannelState",
-  "Navigator",
-]
-version = "0.3.51"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 chrono = "0.4.19"
 
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-js-sys = "0.3.51"
-wasm-bindgen = { version = "0.2.74", features = ["serde-serialize"] }
-wasm-bindgen-futures = "0.4.24"
-wasm-timer = "0.2"
-
 [dev-dependencies]
-simple-logging = "2.0.2"
-tokio = {version = "1", features = ["rt", "macros"]}
-# simple_logger = {version = "1.13.0", features = ["stderr"]}
-simplelog = "0.10"
-anyhow = "1"
-flexi_logger = "0.19"
 env_logger = "0.9.0"
