+++ flnode/src/node/timer.rs
@@ -1,4 +1,4 @@
-use flutils::broker::Broker;
+use flutils::{broker::Broker, schedule::schedule_repeating};
 
 use crate::node::modules::messages::BrokerMessage;
 
@@ -13,30 +13,6 @@ pub enum BrokerTimer {
 pub struct Timer {
     seconds: u32,
     broker: Broker<BrokerMessage>,
-}
-
-#[cfg(target_arch = "wasm32")]
-pub fn schedule_repeating<F>(cb: F)
-where
-    F: 'static + FnMut() + Send,
-{
-    use wasm_bindgen::prelude::*;
-
-    #[wasm_bindgen]
-    extern "C" {
-        pub fn setInterval(callback: JsValue, millis: u32) -> f64;
-    }
-    let ccb = Closure::wrap(Box::new(cb) as Box<dyn FnMut()>);
-    setInterval(ccb.into_js_value(), 1000);
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-pub fn schedule_repeating<F>(cb: F)
-where
-    F: 'static + FnMut() + Send,
-{
-    let timer = timer::Timer::new();
-    timer.schedule_repeating(chrono::Duration::seconds(1), cb);
 }
 
 impl Timer {
