+++ flutils/Cargo.toml
@@ -13,11 +13,14 @@ sha2 = "0.10.1"
 futures = "0.3.19"
 chrono = "0.4.19"
 log = "0.4.14"
-flexi_logger = "0.22.2"
-tokio = {version = "1", features = ["rt", "macros"]}
+timer = "0.2.0"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 js-sys = "0.3.51"
 wasm-bindgen = { version = "0.2.74", features = ["serde-serialize"] }
 wasm-bindgen-futures = "0.4.24"
 wasm-timer = "0.2"
+
+[dev-dependencies]
+tokio = {version = "1", features = ["rt", "macros"]}
+flexi_logger = "0.22.3"
