+++ flutils/src/lib.rs
@@ -2,3 +2,4 @@ pub mod data_storage;
 pub mod nodeids;
 pub mod utils;
 pub mod broker;
+pub mod schedule;
