+++ web/Cargo.toml
@@ -38,6 +38,3 @@ features = [
 [dev-dependencies]
 wasm-bindgen-test = "0.3.24"
 
-[profile.release]
-# Tell `rustc` to optimize for small code size.
-opt-level = "s"
